Refactor ultros crate: extract universalis listener and break down web.rs
- Extract `run_socket_listener` to `ultros/src/universalis_listener.rs`
- Break down `ultros/src/web.rs` into smaller modules:
  - `ultros/src/web/api/retainers.rs`
  - `ultros/src/web/api/lists.rs`
  - `ultros/src/web/api/characters.rs`
  - `ultros/src/web/api/listings.rs`
  - `ultros/src/web/api/user.rs`
  - `ultros/src/web/api/search.rs`
- Extract static file handling to `ultros/src/web/static_files.rs`
- Update `ultros/src/main.rs` and `ultros/src/web.rs` to use new modules.
