+++ ultros/src/web/api/mod.rs
@@ -1,8 +1,14 @@
 mod best_deals;
+pub(crate) mod characters;
 mod cheapest_per_world;
+pub(crate) mod listings;
+pub(crate) mod lists;
 pub(crate) mod real_time_data;
 mod recent_sales;
+pub(crate) mod retainers;
+pub(crate) mod search;
 mod trends;
+pub(crate) mod user;
 
 pub(crate) use best_deals::get_best_deals;
 pub(crate) use cheapest_per_world::cheapest_per_world;
