+++ ultros/src/web/api/search.rs
@@ -1,0 +1,16 @@
+use crate::search_service::SearchService;
+use axum::Json;
+use axum::extract::{Query, State};
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub(crate) struct SearchQuery {
+    q: String,
+}
+
+pub(crate) async fn search(
+    State(service): State<SearchService>,
+    Query(query): Query<SearchQuery>,
+) -> Json<Vec<ultros_api_types::search::SearchResult>> {
+    Json(service.search(&query.q))
+}
